# Improve registry file migration performance